refactor(endpoints): use objects instead of classes
