--- src/core/endpoints/GuideEndpoint.ts
@@ -1,9 +1,0 @@
-/**
- * Represents InnerTube's `/guide` endpoint.
- */
-export default class GuideEndpoint {
-  /**
-   * The endpoint's path.
-   */
-  static PATH: '/guide' = '/guide' as const;
-}
