--- src/core/endpoints/NotificationGetUnseenCountEndpoint.ts
@@ -1,9 +1,0 @@
-/**
- * Represents InnerTube's `/notification/get_unseen_count` endpoint.
- */
-export default class NotificationGetUnseenCountEndpoint {
-  /**
-   * The endpoint's path.
-   */
-  static PATH: '/notification/get_unseen_count' = '/notification/get_unseen_count' as const;
-}
