--- src/core/endpoints/index.ts
@@ -1,18 +1,0 @@
-export * from './BrowseEndpoint.js';
-export { default as BrowseEndpoint } from './BrowseEndpoint.js';
-export * from './GetNotificationMenuEndpoint.js';
-export { default as GetNotificationMenuEndpoint } from './GetNotificationMenuEndpoint.js';
-export * from './GuideEndpoint.js';
-export { default as GuideEndpoint } from './GuideEndpoint.js';
-export * from './MusicGetSearchSuggestionsEndpoint.js';
-export { default as MusicGetSearchSuggestionsEndpoint } from './MusicGetSearchSuggestionsEndpoint.js';
-export * from './NextEndpoint.js';
-export { default as NextEndpoint } from './NextEndpoint.js';
-export * from './NotificationGetUnseenCountEndpoint.js';
-export { default as NotificationGetUnseenCountEndpoint } from './NotificationGetUnseenCountEndpoint.js';
-export * from './PlayerEndpoint.js';
-export { default as PlayerEndpoint } from './PlayerEndpoint.js';
-export * from './ResolveUrlEndpoint.js';
-export { default as ResolveUrlEndpoint } from './ResolveUrlEndpoint.js';
-export * from './SearchEndpoint.js';
-export { default as SearchEndpoint } from './SearchEndpoint.js';
