+++ src/types/index.ts
@@ -1,4 +1,4 @@
 export * from './Cache.js';
-
 export type { default as PlatformShim } from './PlatformShim.js';
 export * from './PlatformShim.js';
+export * from './Endpoints.js';
